backend/requirements.txt
fastapi==0.115.6
instagrapi==2.1.3
pydantic==1.10.12
python-dotenv==1.0.1
